+++ tests/object_array_test.cc
@@ -139,7 +139,7 @@ TEST(ObjectArray, basic) {
   OBJECT_UNREF(obj);
 }
 
-TEST(ObjectArray, dup) {
+TEST(ObjectArray, clone) {
   value_t v;
   string log;
   object_t* dup = NULL;
@@ -149,7 +149,7 @@ TEST(ObjectArray, dup) {
   ASSERT_EQ(object_set_prop(obj, "-1", value_set_int(&v, 2)), RET_OK);
   ASSERT_EQ(object_set_prop(obj, "-1", value_set_int(&v, 3)), RET_OK);
 
-  dup = object_array_clone(OBJECT_ARRAY(obj));
+  dup = object_array_clone(obj);
 
   log = "";
   object_foreach_prop(dup, visit_dump, &log);
@@ -347,9 +347,9 @@ TEST(ObjectArray, get_and_remove) {
   OBJECT_UNREF(obj);
 }
 
-TEST(ObjectArray, create_from_str1) {
+TEST(ObjectArray, create_with_str1) {
   value_t v;
-  object_t* obj = object_array_create_from_str("1,2,3,4", ",", VALUE_TYPE_INT32);
+  object_t* obj = object_array_create_with_str("1,2,3,4", ",", VALUE_TYPE_INT32);
   object_array_t* o = OBJECT_ARRAY(obj);
 
   ASSERT_EQ(o->size, 4);
@@ -368,9 +368,9 @@ TEST(ObjectArray, create_from_str1) {
   OBJECT_UNREF(obj);
 }
 
-TEST(ObjectArray, create_from_str2) {
+TEST(ObjectArray, create_with_str2) {
   value_t v;
-  object_t* obj = object_array_create_from_str("1, 2, 3, 4, ", ", ", VALUE_TYPE_INT32);
+  object_t* obj = object_array_create_with_str("1, 2, 3, 4, ", ", ", VALUE_TYPE_INT32);
   object_array_t* o = OBJECT_ARRAY(obj);
 
   ASSERT_EQ(o->size, 4);
@@ -389,9 +389,9 @@ TEST(ObjectArray, create_from_str2) {
   OBJECT_UNREF(obj);
 }
 
-TEST(ObjectArray, create_from_str3) {
+TEST(ObjectArray, create_with_str3) {
   value_t v;
-  object_t* obj = object_array_create_from_str("1", ", ", VALUE_TYPE_INT32);
+  object_t* obj = object_array_create_with_str("1", ", ", VALUE_TYPE_INT32);
   object_array_t* o = OBJECT_ARRAY(obj);
 
   ASSERT_EQ(o->size, 1);
@@ -401,9 +401,9 @@ TEST(ObjectArray, create_from_str3) {
   OBJECT_UNREF(obj);
 }
 
-TEST(ObjectArray, create_from_str4) {
+TEST(ObjectArray, create_with_str4) {
   value_t v;
-  object_t* obj = object_array_create_from_str("1, ", ", ", VALUE_TYPE_INT32);
+  object_t* obj = object_array_create_with_str("1, ", ", ", VALUE_TYPE_INT32);
   object_array_t* o = OBJECT_ARRAY(obj);
 
   ASSERT_EQ(o->size, 1);
@@ -413,9 +413,9 @@ TEST(ObjectArray, create_from_str4) {
   OBJECT_UNREF(obj);
 }
 
-TEST(ObjectArray, create_from_str5) {
+TEST(ObjectArray, create_with_str5) {
   value_t v;
-  object_t* obj = object_array_create_from_str("abc,123", ",", VALUE_TYPE_STRING);
+  object_t* obj = object_array_create_with_str("abc,123", ",", VALUE_TYPE_STRING);
   object_array_t* o = OBJECT_ARRAY(obj);
 
   ASSERT_EQ(o->size, 2);
@@ -428,3 +428,224 @@ TEST(ObjectArray, create_from_str5) {
   OBJECT_UNREF(obj);
 }
 
+TEST(ObjectArray, join0) {
+  str_t s;
+  object_t* obj = object_array_create_with_str("", ",", VALUE_TYPE_STRING);
+
+  str_init(&s, 100);
+  ASSERT_EQ(object_array_join(obj, ",", &s), RET_OK);
+  ASSERT_STREQ(s.str, "");
+  str_reset(&s);
+  OBJECT_UNREF(obj);
+}
+
+TEST(ObjectArray, join1) {
+  str_t s;
+  object_t* obj = object_array_create_with_str("abc", ",", VALUE_TYPE_STRING);
+
+  str_init(&s, 100);
+  ASSERT_EQ(object_array_join(obj, ",", &s), RET_OK);
+  ASSERT_STREQ(s.str, "abc");
+  str_reset(&s);
+  OBJECT_UNREF(obj);
+}
+
+TEST(ObjectArray, join2) {
+  str_t s;
+  object_t* obj = object_array_create_with_str("a,b,c", ",", VALUE_TYPE_STRING);
+
+  str_init(&s, 100);
+  ASSERT_EQ(object_array_join(obj, "; ", &s), RET_OK);
+  ASSERT_STREQ(s.str, "a; b; c");
+  str_reset(&s);
+  OBJECT_UNREF(obj);
+}
+
+TEST(ObjectArray, dup0) {
+  object_t* dup = NULL;
+  object_t* obj = object_array_create_with_str("", ",", VALUE_TYPE_INT32);
+
+  dup = object_array_dup(obj, 0, 0);
+  ASSERT_EQ(dup != NULL, true);
+  OBJECT_UNREF(dup);
+
+  OBJECT_UNREF(obj);
+}
+
+TEST(ObjectArray, dup1) {
+  value_t v;
+  object_t* dup = NULL;
+  object_t* obj = object_array_create_with_str("1,2,3", ",", VALUE_TYPE_INT32);
+
+  dup = object_array_dup(obj, 0, 1);
+  ASSERT_EQ(dup != NULL, true);
+  ASSERT_EQ(OBJECT_ARRAY(dup)->size, 1);
+  ASSERT_EQ(object_array_get(dup, 0, &v), RET_OK);
+  ASSERT_EQ(value_int(&v), 1);
+  OBJECT_UNREF(dup);
+
+  OBJECT_UNREF(obj);
+}
+
+TEST(ObjectArray, dup2) {
+  value_t v;
+  object_t* dup = NULL;
+  object_t* obj = object_array_create_with_str("1,2,3", ",", VALUE_TYPE_INT32);
+
+  dup = object_array_dup(obj, 0, 2);
+  ASSERT_EQ(dup != NULL, true);
+  ASSERT_EQ(OBJECT_ARRAY(dup)->size, 2);
+  ASSERT_EQ(object_array_get(dup, 0, &v), RET_OK);
+  ASSERT_EQ(value_int(&v), 1);
+  ASSERT_EQ(object_array_get(dup, 1, &v), RET_OK);
+  ASSERT_EQ(value_int(&v), 2);
+  OBJECT_UNREF(dup);
+
+  OBJECT_UNREF(obj);
+}
+
+TEST(ObjectArray, dup3) {
+  value_t v;
+  object_t* dup = NULL;
+  object_t* obj = object_array_create_with_str("1,2,3", ",", VALUE_TYPE_INT32);
+
+  dup = object_array_dup(obj, 1, 3);
+  ASSERT_EQ(dup != NULL, true);
+  ASSERT_EQ(OBJECT_ARRAY(dup)->size, 2);
+  ASSERT_EQ(object_array_get(dup, 0, &v), RET_OK);
+  ASSERT_EQ(value_int(&v), 2);
+  ASSERT_EQ(object_array_get(dup, 1, &v), RET_OK);
+  ASSERT_EQ(value_int(&v), 3);
+  OBJECT_UNREF(dup);
+
+  OBJECT_UNREF(obj);
+}
+
+TEST(ObjectArray, sort_int1) {
+  value_t v;
+  object_t* obj = object_array_create_with_str("2,1,3", ",", VALUE_TYPE_INT32);
+
+  object_array_sort_as_int(obj, TRUE);
+  ASSERT_EQ(object_array_get(obj, 0, &v), RET_OK);
+  ASSERT_EQ(value_int(&v), 1);
+  ASSERT_EQ(object_array_get(obj, 1, &v), RET_OK);
+  ASSERT_EQ(value_int(&v), 2);
+  ASSERT_EQ(object_array_get(obj, 2, &v), RET_OK);
+  ASSERT_EQ(value_int(&v), 3);
+
+  OBJECT_UNREF(obj);
+}
+
+TEST(ObjectArray, sort_int2) {
+  value_t v;
+  object_t* obj = object_array_create_with_str("3,2,1", ",", VALUE_TYPE_INT32);
+
+  object_array_sort_as_int(obj, TRUE);
+  ASSERT_EQ(object_array_get(obj, 0, &v), RET_OK);
+  ASSERT_EQ(value_int(&v), 1);
+  ASSERT_EQ(object_array_get(obj, 1, &v), RET_OK);
+  ASSERT_EQ(value_int(&v), 2);
+  ASSERT_EQ(object_array_get(obj, 2, &v), RET_OK);
+  ASSERT_EQ(value_int(&v), 3);
+
+  OBJECT_UNREF(obj);
+}
+
+TEST(ObjectArray, sort_int3) {
+  value_t v;
+  object_t* obj = object_array_create_with_str("1,2,3", ",", VALUE_TYPE_INT32);
+
+  object_array_sort_as_int(obj, TRUE);
+  ASSERT_EQ(object_array_get(obj, 0, &v), RET_OK);
+  ASSERT_EQ(value_int(&v), 1);
+  ASSERT_EQ(object_array_get(obj, 1, &v), RET_OK);
+  ASSERT_EQ(value_int(&v), 2);
+  ASSERT_EQ(object_array_get(obj, 2, &v), RET_OK);
+  ASSERT_EQ(value_int(&v), 3);
+
+  OBJECT_UNREF(obj);
+}
+
+TEST(ObjectArray, sort_int4) {
+  value_t v;
+  object_t* obj = object_array_create_with_str("1,2,3", ",", VALUE_TYPE_INT32);
+
+  object_array_sort_as_int(obj, FALSE);
+  ASSERT_EQ(object_array_get(obj, 0, &v), RET_OK);
+  ASSERT_EQ(value_int(&v), 3);
+  ASSERT_EQ(object_array_get(obj, 1, &v), RET_OK);
+  ASSERT_EQ(value_int(&v), 2);
+  ASSERT_EQ(object_array_get(obj, 2, &v), RET_OK);
+  ASSERT_EQ(value_int(&v), 1);
+
+  OBJECT_UNREF(obj);
+}
+
+TEST(ObjectArray, sort_double1) {
+  value_t v;
+  object_t* obj = object_array_create_with_str("1,2,3", ",", VALUE_TYPE_DOUBLE);
+
+  object_array_sort_as_double(obj, TRUE);
+  ASSERT_EQ(object_array_get(obj, 0, &v), RET_OK);
+  ASSERT_EQ(value_int(&v), 1);
+  ASSERT_EQ(object_array_get(obj, 1, &v), RET_OK);
+  ASSERT_EQ(value_int(&v), 2);
+  ASSERT_EQ(object_array_get(obj, 2, &v), RET_OK);
+  ASSERT_EQ(value_int(&v), 3);
+
+  OBJECT_UNREF(obj);
+}
+
+TEST(ObjectArray, sort_double2) {
+  value_t v;
+  object_t* obj = object_array_create_with_str("1,2,3", ",", VALUE_TYPE_DOUBLE);
+
+  object_array_sort_as_double(obj, FALSE);
+  ASSERT_EQ(object_array_get(obj, 0, &v), RET_OK);
+  ASSERT_EQ(value_int(&v), 3);
+  ASSERT_EQ(object_array_get(obj, 1, &v), RET_OK);
+  ASSERT_EQ(value_int(&v), 2);
+  ASSERT_EQ(object_array_get(obj, 2, &v), RET_OK);
+  ASSERT_EQ(value_int(&v), 1);
+
+  OBJECT_UNREF(obj);
+}
+
+TEST(ObjectArray, sort_str1) {
+  value_t v;
+  object_t* obj = object_array_create_with_str("a1,B2,c3", ",", VALUE_TYPE_STRING);
+
+  object_array_sort_as_str(obj, TRUE, FALSE);
+  ASSERT_EQ(object_array_get(obj, 0, &v), RET_OK);
+  ASSERT_STREQ(value_str(&v), "B2");
+  ASSERT_EQ(object_array_get(obj, 1, &v), RET_OK);
+  ASSERT_STREQ(value_str(&v), "a1");
+
+  OBJECT_UNREF(obj);
+}
+
+TEST(ObjectArray, sort_str2) {
+  value_t v;
+  object_t* obj = object_array_create_with_str("a1,B2,c3", ",", VALUE_TYPE_STRING);
+
+  object_array_sort_as_str(obj, TRUE, TRUE);
+  ASSERT_EQ(object_array_get(obj, 0, &v), RET_OK);
+  ASSERT_STREQ(value_str(&v), "a1");
+  ASSERT_EQ(object_array_get(obj, 1, &v), RET_OK);
+  ASSERT_STREQ(value_str(&v), "B2");
+
+  OBJECT_UNREF(obj);
+}
+
+TEST(ObjectArray, sort_str3) {
+  value_t v;
+  object_t* obj = object_array_create_with_str("a1,B2,c3", ",", VALUE_TYPE_STRING);
+
+  object_array_sort_as_str(obj, FALSE, TRUE);
+  ASSERT_EQ(object_array_get(obj, 0, &v), RET_OK);
+  ASSERT_STREQ(value_str(&v), "c3");
+  ASSERT_EQ(object_array_get(obj, 1, &v), RET_OK);
+  ASSERT_STREQ(value_str(&v), "B2");
+
+  OBJECT_UNREF(obj);
+}